+++ CHANGES.rst
@@ -12,6 +12,8 @@
 - Overhaul documentation so that all essential functionality is presented on
   the front page.
 
+- Move ``ligo.skymap.command`` to top-level ``ligo.skymap.tool`` module.
+
 0.0.6 (2018-04-13)
 ==================